# arm/upper_computer/gateway/openapi/runtime_api.yaml
...
openapi: 3.1.0
info:
  title: Embodied Arm HMI Gateway
  version: 2.3.0
paths:
...
  /api/task/start:
    post:
      summary: Post Task Start
      operationId: post_task_start_api_task_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartTaskRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartTaskResponseEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/hardware/set-mode:
    post:
      summary: Post Set Mode
      description: "Update the controller mode through the hardware command surface.\n\
        \nArgs:\n    body: Requested controller mode payload.\n    request: FastAPI\
        \ request carrying runtime context and operator headers.\n\nReturns:\n   \
        \ Wrapped command result bound to the current request id.\n\nRaises:\n   \
        \ ApiException: Propagates role or transport failures through the command\
        \ pipeline.\n\nBoundary behavior:\n    Operator role may switch only into\
        \ public operator modes; maintenance-only\n    modes still require the maintainer\
        \ role and never bypass readiness fail-closed\n    handling on downstream\
        \ commands."
      operationId: post_set_mode_api_hardware_set_mode_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetModeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/hardware/gripper:
    post:
      summary: Post Gripper
      operationId: post_gripper_api_hardware_gripper_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GripperRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/hardware/jog-joint:
    post:
      summary: Post Jog Joint
      operationId: post_jog_joint_api_hardware_jog_joint_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JogJointRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/hardware/servo-cartesian:
    post:
      summary: Post Servo Cartesian
      description: "Dispatch a validated cartesian servo command to the ROS bridge.\n\
        \nArgs:\n    body: Validated cartesian servo command payload.\n    request:\
        \ FastAPI request carrying the runtime app context and actor headers.\n\n\
        Returns:\n    A wrapped command-acceptance response bound to the current request\
        \ id.\n\nRaises:\n    ApiException: Raised when role or readiness policy blocks\
        \ the command.\n\nBoundary behavior:\n    Returns HTTP 501 for intentionally\
        \ disabled contract branches and HTTP 409\n    for readiness-gated blocks\
        \ without mutating gateway state."
      operationId: post_servo_cartesian_api_hardware_servo_cartesian_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServoCartesianRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GripperRequest:
      properties:
        open:
          type: boolean
          title: Open
          default: true
      type: object
      title: GripperRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JogJointRequest:
      properties:
        jointIndex:
          type: integer
          maximum: 5.0
          minimum: 0.0
          title: Jointindex
        direction:
          type: integer
          enum:
          - -1
          - 1
          title: Direction
          default: 1
        stepDeg:
          type: number
          maximum: 10.0
          exclusiveMinimum: 0.0
          title: Stepdeg
          default: 2.0
      type: object
      required:
      - jointIndex
      title: JogJointRequest
    ServoCartesianRequest:
      properties:
        axis:
          type: string
          enum:
          - x
          - y
          - z
          - rx
          - ry
          - rz
          title: Axis
        delta:
          type: number
          maximum: 0.1
          minimum: -0.1
          title: Delta
      type: object
      required:
      - axis
      - delta
      title: ServoCartesianRequest
    SetModeRequest:
      properties:
        mode:
          type: string
          enum:
          - idle
          - manual
          - task
          - maintenance
          title: Mode
      type: object
      required:
      - mode
      title: SetModeRequest
    StartTaskDecisionResponse:
      properties:
        taskId:
          type: string
          title: Taskid
        taskRunId:
          type: string
          title: Taskrunid
        episodeId:
          type: string
          title: Episodeid
        templateId:
          type: string
          title: Templateid
        pluginKey:
          type: string
          title: Pluginkey
        graphKey:
          type: string
          title: Graphkey
        runtimeTier:
          type: string
          enum:
          - preview
          - validated_sim
          - validated_live
          title: Runtimetier
        productLine:
          type: string
          title: Productline
        catalogSchemaVersion:
          type: integer
          title: Catalogschemaversion
        templateVersion:
          type: integer
          title: Templateversion
        pluginContractVersion:
          type: integer
          title: Plugincontractversion
        success:
          type: boolean
          title: Success
          default: true
        accepted:
          type: boolean
          title: Accepted
          default: true
        authoritativeStatus:
          anyOf:
          - type: string
            enum:
            - accepted
            - success
            - failed
            - blocked
            - rejected
            - observed
          - type: string
          title: Authoritativestatus
          default: accepted
        completionPending:
          type: boolean
          title: Completionpending
          default: true
        localPreviewOnly:
          type: boolean
          title: Localpreviewonly
          default: false
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        operationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Operationid
        requestId:
          anyOf:
          - type: string
          - type: 'null'
          title: Requestid
        receiptId:
          anyOf:
          - type: string
          - type: 'null'
          title: Receiptid
        correlationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlationid
      type: object
      required:
      - taskId
      - taskRunId
      - episodeId
      - templateId
      - pluginKey
      - graphKey
      - runtimeTier
      - productLine
      - catalogSchemaVersion
      - templateVersion
      - pluginContractVersion
      title: StartTaskDecisionResponse
    StartTaskRequest:
      properties:
        templateId:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Templateid
        taskType:
          type: string
          enum:
          - pick_place
          - sort_by_color
          - sort_by_qr
          - clear_table
          title: Tasktype
          default: pick_place
        targetCategory:
          anyOf:
          - type: string
            enum:
            - red
            - blue
            - green
            - qr_a
            - qr_b
            - qr_c
          - type: 'null'
          title: Targetcategory
      type: object
      title: StartTaskRequest
    StartTaskResponseEnvelope:
      properties:
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: ok
        requestId:
          type: string
          title: Requestid
        correlationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlationid
        timestamp:
          type: string
          title: Timestamp
        data:
          $ref: '#/components/schemas/StartTaskDecisionResponse'
      type: object
      required:
      - requestId
      - timestamp
      - data
      title: StartTaskResponseEnvelope
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError